Triangle phase, initial pop fix, noise blip quality
Inverted the triangle phase to match that of the hardware, making volume of DPCM and Noise much closer match (notably if the song has no triangle).

Worked around an initial pop caused by this, by using an offset, and also skip the very first sample to eliminate a tiny remaining pop.

Increased the noise blip to good, as it really seems to much closer match tests compared to the real hardware. Much higher frequencies are playing on the real hardware.