Workaround for extremely short midi notes
The stop note would occur on the same frame and cancel out the note with a stop note. This is a bit of a hack, but it makes it so the stop note pushes forward by a frame. Another option that worked was to set the duration to 1 for the last note, then convert to simple notes and then convert to compound notes afterwards, but that's probably even more tedious than this.