Rework C bindings to reduce overhead
Removes the C specific entrypoints in the ca65 version of famistudio
in favor of using custom macros to adapt to the FS calling convention.
This reduces the work needed to keep parity between the different assemblers,
but adds more complexity to the C header file itself.

This allows us to also add a similar header file for LLVM-MOS, which is
able to link directly with CA65 object files. There are differences in the
two compiler's calling conventions, so this allows us to reuse the same
assembler file, and adapt the calling convention only in the C header file.

Lastly, its seemingly not possible to do these same changes to the SDCC build
so we leave the CFG_C_BINDINGS option in there in order to keep build parity
when testing with the Unit Tests.