Attempt to fix crash for paths without read access
Linux doesn't work with try / catch, so it has to be done on an OS level. MacOS uses OS dialogs, but the code is still there just to unify desktop platform code. Windows still uses try / catch to test for read access. as it doesn't use libc. This was done since Windows has the option to use native dialogs.