Cleanup
GNOME is not as ideal, but it seems that it's just designed that way in general in terms of keeping windows on top etc. It often sends a notification rather than prompting, and sometimes dialogs render in the upper left.

Wayland still seems to have intermittent problems for me when using vscode, but only on Arch Linux. Building the program and launching it works fine, and vscode seems to be fine on Debian 12 also.