FDS Modulation improvements / fixes
How2Bboss had pointed out NSF imports that had multiwave with automod weren't working properly. This was a result of us writing the mod table every time a new instrument is used, even when it used the same envelope. This fixes the issue and also optimises the ASM code, since it doesn't have to keep needlessly reloading the mod table every instrument change.