Regional numeric fix
Some languages were parsing the wrong "-" symbol. This caused a crash in the mixer settings, and a lack of negative numeric values in NumericUpDown.

No idea why OnRender for Slider was being converted to a string (from an int) and then to a float, that must have been a mistake.